AG-9017 Add new file errorBarNode.ts
The ErrorBarPoint and ErrorBarNode types are being moved into this file
because they will likely become much larger in implementation size.
