+++ packages/ag-charts-enterprise/src/error-bar/errorBar.ts
@@ -1,5 +1,6 @@
 import type { _Scale } from 'ag-charts-community';
 import { _Scene, _Util, _ModuleSupport } from 'ag-charts-community';
+import { ErrorBarPoints, ErrorBarNode } from './errorBarNode';
 
 const { ChartAxisDirection, Validate, OPT_STRING } = _ModuleSupport;
 
@@ -7,32 +8,6 @@ type CartesianSeries<
     C extends _ModuleSupport.SeriesNodeDataContext<any, any>,
     N extends _Scene.Node
 > = _ModuleSupport.CartesianSeries<C, N>;
-
-export interface ErrorBarPoints {
-    readonly yLowerPoint: _Scene.Point;
-    readonly yUpperPoint: _Scene.Point;
-}
-
-export class ErrorBarNode extends _Scene.Path {
-    public points: ErrorBarPoints = { yLowerPoint: { x: 0, y: 0 }, yUpperPoint: { x: 0, y: 0 } };
-
-    updatePath() {
-        // TODO(olegat) implement, this is just placeholder draw code
-        const { path } = this;
-        this.fill = 'black';
-        this.stroke = 'black';
-        this.strokeWidth = 1;
-        this.fillOpacity = 1;
-        this.strokeOpacity = 1;
-
-        const { yLowerPoint, yUpperPoint } = this.points;
-
-        path.clear();
-        path.moveTo(yLowerPoint.x, yLowerPoint.y);
-        path.lineTo(yUpperPoint.x, yUpperPoint.y);
-        path.closePath();
-    }
-}
 
 export class ErrorBars extends _ModuleSupport.BaseModuleInstance implements _ModuleSupport.ModuleInstance {
     @Validate(OPT_STRING)
